% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Load FARS data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector or a list with years}
}
\value{
a list with data tibbles year by year
}
\description{
This function loads data from FARS data files and 
puts them in a list of tibbles, year by year.
}
\note{
Missing data for any of the requested years will
result in a warning and a \code{NULL} in a coresponding
place in a return list. Set the working directory before running
this function.
}
\examples{
# good years, but bad path
fars_from_2013_to_2015 <- fars_read_years(2013:2015)
fars_from_2013_to_2015
# the directory with data files must be a working directory
setwd("data")
fars_from_2013_to_2015 <- fars_read_years(2013:2015)
str(fars_from_2013_to_2015)
# invalid range results in NULLs
fars_from_2010_to_2013 <- fars_read_years(2010:2013)
str(fars_from_2013_to_2015)
}
