% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count observations from FARS data files}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector or a list with years}
}
\value{
a tibble with numbers of observations in months
}
\description{
This function loads FARS data fiels for requested years,
calculates observations per month in every year and returns 
the result in a tibble
}
\note{
The function will return an error if no valid year
is given in the years parameter
}
\examples{
# full years range is covered in data
fars_summarize_years(2013:2015)
# no year is covered
fars_summarize_years(2011:2012)
# invalid years are removed
fars_summarize_years(2011:2014)
}
