% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a file name for FARS annual data}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numeric or character with a year number}
}
\value{
a character with the full file name
}
\description{
This function creates a file name of the structure: 
\code{accident_[YEAR].csv.bz2} given the year.
}
\note{
This function will return \code{"accident_NA.csv.bz2"} and 
 throw a warning if the year argument cannot be transoformed to integer
}
\examples{
# acceptable input
make_filename(2015)
make_filename("2015")
# wrong input
make_filename("two thousand fifteen")
}
